Handle lvgl.h like in lv_drivers, not in #if LV_CONF_INCLUDE_SIMPLE
Include of lvgl main header has nothing to do with the configuration
include; hence it should not be part of `LV_CONF_INCLUDE_SIMPLE`
preprocessor directive. Handle it alike in lv_drivers, with

    #ifdef LV_LVGL_H_INCLUDE_SIMPLE
    #include "lvgl.h"
    #else
    #include "lvgl/lvgl.h"
    #endif

in `lv_examples.h`, which is then included on top of each header file in
lv_examples folders. In the end, `lvgl.h` must be included before the
`LV_USE_TESTS` preprocessor directive because, for some tests, there is
also a condition on a lvgl feature, like `LV_USE_GROUP` which is defined
by the main configuration file `lv_conf.h` which is included by
`lvgl.h`.